refactor: address remaining items from PR review
Three cleanup findings from the prior audit that were initially deferred:

- Extract the OPTIONAL_FIELD_CLAMPS table to shared/timeline/item-clamps so
  the runtime items-store normalizer and the project-load migration share
  one source of truth; collapses ~150 lines of repeated if-clamp branches
  in migrations/normalize.ts.
- Move replaceableCaptionClipIds out of withItemIndexes' hot path: it now
  rebuilds lazily via selectReplaceableCaptionClipIds, keyed on items
  identity, so per-frame drag mutations skip the O(N) legacy-caption
  detection pass.
- Drop the redundant top-level xxxRef.current = xxx writes in
  use-smart-trim-hover; the syncXxx helpers were already updating the
  refs, the duplicate writes only existed to catch the tool-change
  useEffect's direct setX calls which now route through syncXxx.
